+++ examples/manager-and-workers/versions.tf
@@ -3,7 +3,12 @@ terraform {
   required_version = ">= 0.13"
   required_providers {
     aws = {
-      source = "hashicorp/aws"
+      source  = "hashicorp/aws"
+      version = ">=5.0.0"
+    }
+    tls = {
+      source  = "hashicorp/tls"
+      version = ">= 4.0.3, < 5.0.0"
     }
   }
 }